[#30] Rename: comment 관련 컴포넌트 comments 폴더로 이동
